backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
asyncpg==0.29.0
pydantic[email]==2.8.2
python-dotenv==1.0.1